+++ src/main/java/com/yh/web/dto/board/BoardFile.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class BoardFile {
-    private int articleNo;
+    private long articleNo;
     private String fileName;
     private long fileSize;
     private String originalFileName;
